TINYINT not supported on pg :wrench:

+++ server/common/database/migrations/20181006191154-ungraded-event.js
@@ -25,7 +25,7 @@ module.exports = {
       allowNull: false
     },
     progress: {
-      type: Sequelize.TINYINT,
+      type: Sequelize.SMALLINT,
       defaultValue: 100,
       allowNull: false
     },
